fix: remove deprecated sudt script from script list
This commit removes an incorrectly recorded sudt script from
script list.

The first version of sudt script is of `hash type: data` but
it was recorded as of `hash type: type`, which is incorrect.

Besides, the first version of sudt script was never committed
into RFCs(there was a draft but closed), so it's not necessary
to be listed.

Ref:
- draft PR mentioned the deprecated sudt script: https://github.com/nervosnetwork/rfcs/pull/199
- PR added the incorrect sudt info: https://github.com/nervosnetwork/ckb-explorer-frontend/commit/0b7e28514fb9697e95aeb25c3d221d605a2573be#diff-5636bb1b68c18c68892b6c62dfc3f46f9af26639e920bf49902e70156eedc2a1R156