fix: fix page crash on clicking capacity tab in cell info dialog
This PR fixed the page crash by clicking the capacity tab in
the cell info dialog. The crash was caused by incorrectly formatted
"declared capacity". It was a BigNumber, which is not a valid
ReactNode.

Besides, the tab index was reset to 0 when request failed, which
made the tab display incorrectly. This issue is also fixed in
this PR